backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
sqlalchemy==2.0.40
pydantic==2.11.3
pydantic-settings==2.8.1
pypdf==5.4.0
python-docx==1.1.2
httpx==0.28.1
python-dotenv==1.1.0
